Split `lib_regions_of_interest/v1.py` into `v1.py` and `v1_checks.py`
